aireplay-ng: Fix unintended sign extension. (CID #1309352)

+++ src/aireplay-ng.c
@@ -3504,8 +3504,8 @@ static int do_attack_chopchop(void)
 				break;
 		}
 
-		crc_mask
-			= crc_tbl[crc_mask & 0xFF] ^ (crc_mask >> 8) ^ (chopped[i] << 24);
+		crc_mask = crc_tbl[crc_mask & 0xFF] ^ (crc_mask >> 8UL)
+				   ^ ((unsigned long) chopped[i] << 24UL);
 	}
 
 	for (i = 0; i < 4; i++)
